Housekeeping of utils and JSDocs (#26)
* Remove unused `utils.multiplyBy()` function

* Fix theoretical max size of a Discord Snowflake
Snowflakes in Discord represent 64 bit numbers. The maximum unsigned 64 bit integer is 18446744073709552000, which as text contains 20 characters, more than 18.

This fixes the check in `util.isSnowflake()` and changes the types of database columns that previously assumed Discord IDs were at most 18 characters long.

* Rework `utils.findX()` functions
- Add regex constants for user, role, and channel mentions
- Add `resolveXID()` functions which extract the IDs out of mentions
- Make all `findX()` functions take in a "manager" for retrieving the X object instead of requiring a message object
- Allow all `findX()` functions to extract the object ID out of mentions using their respective `resolveXID()` functions

* Make config extension use `utils.findGuildChannel()`

* housekeeping: Remove type information from JSDocs
The type is already available through TypeScript, so having it in the JSDocs was just clutter

* Fix config's `objectToString()` taking in the database value instead of the js value when executed through !config
This was causing, for example, the starboard_channel to appear as #undefined, since `objectToString()` was receiving the channel's ID instead of the channel object.